Fix tests for responsive util from create-styles
