+++ packages/create-styles/src/createCompiler/index.js
@@ -1,2 +1,3 @@
 export * from './createCompiler';
 export * from './createCSS';
+export * from './responsive';
